fix: delete media management

+++ src/main/java/techquack/com/onestar3gram/services/storage/FileSystemStorageService.java
@@ -91,6 +91,10 @@ public class FileSystemStorageService implements StorageService {
         if (!mediaFileRepository.existsById(id)) {
             throw new FileNotFoundException("File not found");
         }
+        File file = getFile(id);
+        if(!file.delete()) {
+            throw new FileNotFoundException("Impossible to delete");
+        }
         mediaFileRepository.deleteById(id);
     }
